consolidated analysis of reliable area activations on the second level
The new ipython notebook provides a more mature analysis of area activations.
Specifically, I now directly estimate the p-values of temporal clusters which I
then subject to FDR-correction. The cluster-defining threshold is still the 95%
quantile of the distribution of values of the chosen measure across all
regressors for which clusters should be identified. This procedure returns 2568
temporal clusters of which 420 survive FDR in the end.

Note that I chose a particularly conservative cluster alpha-level of 0.001 in
the FDR-correction, because I had the impression that lower values still select
a few spurious/hard to interpret clusters (with alpha=0.01 795 clusters are
identified instead of 420), for example, a series of brief activations for dot_x
wandering from medial to lateral temporal cortex to the posterior cingulate to
lateral temporal cortex to insular and auditory cortex starting straight with
dot onset. It may be that this is genuine and it may be not. So I decided to go
with a more conservative threshold.

I still find it hard to consolidate all the possible findings, but it appears
quite informative to contrast activations at different time points. For example,
for dot_x the difference between time points at 120 and 170 ms is that the early
peak shows more activations in early visual cortex together with temporal areas
whereas in the later peak dorsal stream visual cortex, parietal cortex and the
posterior cingulate are engaged while area v23ab and left V2 are active at both
time points. This is perhaps what you would expect from the visual signal
climbing up through the hierarchy. So I wonder what I can learn additionally
from that?

Similarly, contrasting effects at 170 ms and 320 ms reveals that effects in
parietal areas at the earlier time point are replaced with effects in motor
areas with a possible bridge in cingulate and visual areas (specifically when
also checking effects at 250 ms). Contrasting 320 ms to 410 ms is a bit harder
with effects at the later time point, but not in the earlier, for areas in
parietal areas previously seen at 170 ms, but also some temporal and motor areas
and a little bit of anterior cingulate and dlPFC activations.