new analysis based on maximum source current in areas
As I cannot say much about the magnitude of effects, because I average across
sources in an area, areas have different sizes and may only be partially
activated, I repeated the analysis with taking the peak values from an area
instead of averaging them. I thought that this would let me observe the actual
magnitude of an effect within the area, but this was misconceived. I thought
that I would select the maximum effect (correlation with regressor) within an
area, but I extract area values before statistical inference. Thus, taking the
maximum source current within an area may select very different effects from
different sub-parts of an area.

Consequently, most previous effects are gone, especially those of dot_x for
which only 11 clusters remain after FDR correction. Interestingly, dot_y still
shows 110 clusters, but these clusters are based on mu_testval, i.e., looking
for consistent effects no matter how big they are. As I specifically did this
analysis to investigate the magnitude of effects, I should check mu_p_large
instead, i.e., look for clusters which reliably show large (>=0.02) effects. In
this analysis dot_x has 22 clusters and dot_y only 8 showing that the dot_y
effects tend to be very small. The found clusters also span only small time
intervals leading to relatively high p-values and few overlapping clusters.

In total the effects are fewer and appear to be less reliable. This said, they
still make a bit of sense. For example, for abs_dot_x, which has still 51
clusters with mu_p_large, many effects are around 140 ms and these are
predominantly in visual cortex while at a later time point around 560 ms
abs_dot_x appears to correlate with the maximum source current in predominantly
frontal areas.

Why do I see 'significant' clusters at all and what does it actually mean, if I
see correlation of the maximum source current in an area with my regressors? If
you assume that many things can be going on in the same area at the same time,
correlation with a variable of interest across trials means that somehow my
variable of interest is represented in that part of the area that currently has
the largest source current for whatever reason. That there is no good
correspondence between largest source current and the variable of interest is
reflected in the small correlations that I find. It is also interesting that I
seem to see more effects in frontal areas with this analysis. I wonder whether
this has something to do with the size of the areas. A quick check appears to
confirm that it is mostly small areas that show effects of dot_x, but then why,
e.g., does area v23ab not turn up anymore? It had pretty strong effects when
averaging within areas and has only 414 vertices which is within the smallest 70
of 360 areas?