determining correlations between dot_x and response for different dots
In plots of average magnitude of correlation with response I noticed that there
are small peaks which could correspond to dot_x effects as their timing is at
520 and 580 ms which are the times of the 120 and 180 ms effect of the 5th dot.
So I wanted to know whether it is possible that response also shows dot_x
effects. To do this I checked the correlations of the corresponding regressors.
Indeed this is highest for the 5th dot with a mean of 0.38 across subjects. As
the potential 5th dot peaks in the response trace are small compared to the main
effect around 800 ms I would argue that this effect is not simply a copy of the
5th dot 300 ms effect despite the similar sensor topology, although it may
contribute.