new analysis on source points of given brain areas (V1, V2)
I used this to extract source currents for dipoles only in V1 and V2
without averaging across the dipoles. This is for an analysis in which I try to
see whether dot_y has early (120 ms) effects, if I do not average across all
dipoles of an area which could favour dot_x as I'm not averaging across the left
and right hemisphere.

resulting source signals can be found in:
	mne_subjects/fsaverage/bem/source_epochs_allsubs_HCPMMP1_201708151431.h5

inference results for GLM directly on dipole currents of V1 and V2 are in:
	source_sequential_201708151458.h5

and the results of the second level folded normal mixture analysis are in the
corresponding files.

Note that I only computed results for the first 200 ms after dot onset (for the
first three dots), but all other options should be the same as in the original
analysis.