difference between dot_x and dot_y in V1, V2 dipoles
The new figure is supposed to summarise my findings for the V1, V2 current
dipoles. Despite the higher spatial resolution the results mirror those of the
average area effects: dot_x shows an early peak around 120 ms while dot_y shows
a flat peak around 170 ms with posterior mu starting to rise only from 120 ms. I
see this clearly when averaging effects across dipoles, as expected, but also
when I check the maximum effects across dipoles. This means that there really
are no dipoles in V1 or V2 which exhibit a noticable correlation with dot_y at
120 ms.

This is good for us as it supports our interpretation that dot_x exhibits early
effects, because it's decision relevant and not because horizontal locations are
preferentially processed by the brain. The argument is that early visual cortex
is not only separated in a left and right hemisphere region which preferentially
processes visual stimuli from the contralateral visual field, but it also
spatially separates in a bottom and top region which preferentially process
stimuli in the opposite visual field. If we see a strong early effect for dot_x,
because of the separation in left and right early visual cortex, we should also
see a similarly strong early effect for dot_y, because of the separation in top
and bottom early visual cortex - provided that our spatial resolution is good
enough that we see separate signals from those left, right and top, bottom areas
of visual cortex.

With my single dipole analysis I tried to do this. The source space I chose
defines dipoles which are only spaced ca. 5 mm from each other which could be
sufficient for this purpose, but I'm not sure whether the MEG signal can
actually discern signals on that scale. It is clear, though, that left and right
visual cortex - at least the lateral parts - are sufficiently far from each
other so that MEG can pick up on that. A further complication is the folding of
the cortex which could mean that certain dipoles may be closer than suggested by
their spacing on the cortical surface.

A short check of results from retinotopy in humans with fMRI suggests that
representations of the top and bottom of the visual field in early visual cortex
are separated by at least about 1 cm on the flattened surface. This is mainly
based on [S A Engel, G H Glover, B A Wandell; Retinotopic organization in human
visual cortex and the spatial precision of functional MRI., Cerebral Cortex,
Volume 7, Issue 2, 1 March 1997, Pages 181–192,
https://doi.org/10.1093/cercor/7.2.181]. Note that apparently the border between
V1 and V2 is also defined by the response to stimuli in the top or bottom of the
visual field. Specifically, the upper border of V1 to V2 is where particularly
stimuli in the bottom visual field lead to a high response and the other way
around for the lower border of V1 to V2. The retinotopy of this paper also
suggests, however, that simply ventral (inferior, lower) parts of early visual
cortex preferrably respond to stimuli in upper visual field while dorsal
(superior, upper) parts of visual cortex prefer stimuli in the lower visual
field. The trouble is that V2 and V3 loop around V1 such that there is a ventral
and dorsal part of both over which I average when extracting mean source
currents for the brain areas. This can all also be seen in Glasser2016, Fig. 3
of the appendix (p. 95 in pdf). This means that we really may have a problem
identifying dot_y correlations on averaged source currents, but that it should
indeed be possible to pick up the effect for non-averaged dipole signals,
because it is sufficiently far spread between ventral and dorsal dipoles.

In conclusion, not seeing that dot_y has an early effect in any of the source
dipoles supports that the reason for the early dot_x effect is not the basic
retinotopic organisation of the cortex, because then we should see the
equivalent effect for dot_y. Given that we don't see a dot_y effect in sensor
space either, we probably shouldn't be surprised that it is not there in source
space. So, still the remaining question is whether MEG would actually be at all
able to pick up the vertical retinotopic effect?

A quick search for retinotopic mapping with MEG brought me to:

Sharon, D., Hämäläinen, M. S., Tootell, R. B. H., Halgren, E., & Belliveau, J.
W. (2007). The advantage of combining MEG and EEG: comparison to fMRI in focally
stimulated visual cortex. NeuroImage, 36(4), 1225–1235.
https://doi.org/10.1016/j.neuroimage.2007.03.066

Their results show that it is possible to reconstruct sources of stimuli
presented at different locations in the visual field with MEG. Specifically, it
should be possible to see a difference between stimuli in the lower and upper
visual field. Also note that the visual evoked response peaks already around
80 ms and that the results are based on the MEG signal at that time point. This
indicates that I do not see the basic visual response in my analysis.
Alternatively, why do I not see a basic visual response at 80 ms?