checking the magnetic fields evoked by dots
Apparently there is an early evoked response in the EEG/MEG signal that is
supposedly related to an early visual response in V1. This evoked response is
supposed to peak between 70 and 80 ms after onset of a visual stimulus. I wanted
to check whether I can detect that response in my data. In the next step I would
then try to see whether I can discriminate top vs. bottom dots in this response
in source space.

I have the impression that I do not see the visual evoked response that has
previously been described. The signal evoked by the first dot exhibits a
relatively broad positive peak around 70-80 ms over an occipital magnetometer
(MEG2111) which could correspond to the previously described visual evoked
potential, but the peak is rather shallow (compared to the following peak around
170-180 ms) and the topology at 50 or 80 ms suggests that the greatest signal
is due to a central-frontal dipole with a peak positivity close to the left ear
and peak negativity close to the right ear. I wonder what this dipole
corresponds to, i.e., what it does. It also exists, very weakly, from dot onset
despite baseline correction for mean signal in -300 to 0 ms. So it must be some
kind of expectation signal that is developing in the baseline period. Perhaps I
should check this, as it could also be interesting on its own.

Note that our task does not allow to observe a pure, traditional visually evoked
response. This is because the appearance of the first dot happens simultaneously
with the disappearance of the fixation cross. So the dot-evoked signal in visual
cortex is mixed with a response to the disappearance of a foveal stimulus,
whatever this may be (probably involving reduction in source currents
somewhere). Additionally, the evoked response for dots following the first is a
mixed signal of the time-shifted evoked responses of the current and all
previous dots which cannot be unmixed by averaging.

In conclusion, perhaps there is a typical, but weak visual evoked response in
the magnetometer signals, but it's certainly mixed with other signals. I want to
check next whether I see this also at the source level in V1 and V2. If I see a
peak there around 80 ms in the overall average, I may also find
location-specific responses at this time point.