VEF not equal to intercept in sensor space because of normalisation
Out of curiosity I wanted to check whether the visually evoked field that I
computed by simple averaging across trials and subjects is equal to the
intercept beta of the singledot regression. Intuitively this should be the case
as the intercept represents the baseline value after accounting for the effects
of the other regressors. I found, however, that the topographies of the VEF and
the singledot intercept at 120 ms after first dot onset do not look the same:
They share the lateral activations near the ears, but the intercept topography
shows additional peaks in the centre. It is, therefore, a bit more similar to
the topography of the VEF when averaging over dots (dot onset aligned epochs
instead of first dot onset aligned).

I have confirmed that this comes from the normalisation of the signal that I do
before the regression. The sensors differ in their overall activation level
across all subjects, trials and times such that the topography of the mean
obviously changes when I subtract this overall activation level. If I normalise
the sensor signals before computing the VEF I get very similar topographies for
the VEF and intercept beta at 120 ms after first dot onset. This means that I
cannot interpret the intercept beta as a VEF, if I normalise the data before
regression.