VEFs, retinotopy, source flipping and semi-transparent, divergent colormaps
After observing at least small visually evoked fields around 70/80 ms after
first dot onset on the sensor level I wanted to see whether I can find these on
the source level. Specifically, I then wanted to check whether I could find
retinotopic effects in early visual cortex that corresond to previous findings
which state that early visual cortex has strong retinotopic organisation with
visual stimuli being preferentially processed in the contralateral hemisphere to
their location on the screen. Similarly V1 clearly divides into a part
preferentially processing stimuli in the upper visual field and a part
processing stimuli in the lower visual field.

See also VEFs_email_to_Stefan_20170821.odt for an overview and discussion of
results.

I did not find any consistent signals in V1 or V2 around 70/80 ms after first
dot onset, i.e., I cannot find the typical VEF in early visual cortex sources.
Specifically, the maximum magnitude of source currents is pretty flat until
about 100 ms.

In the next step I built a basic regression analysis in which I classify the
locations of presented dots into left vs. right and top vs. bottom to see
whether source currents vary according to this classification. I excluded dots
from analysis which were within 50 px from the centre of the screen (along
either axis). This is supposed to exclude foveal stimulation from analysis as I
expect clearer signals from the more distant dots, but the threshold of 50 px is
based on pure intuition. I only look at group-level results and get these by
simply applying a t-test to contrasts of interest. Apart from the 4 dot
regressors these are the vertical contrast (top - bottom) and the horizontal
contrast (left - right).

I checked the t-maps for the horizontal and vertical contrasts at 80 ms, but
couldn't recognise any patterns. As the VEF analysis suggested, there doesn't
appear to be anything interesting in early visual cortex before 100 ms in our
data. Interesting patterns become visible, however, at the later time points.
For example, I see a clear lateralisation for the horizontal contrast at 120 and
180 ms. Even more intresting, for the vertical contrast at 180 ms V1 is divided
in both hemispheres along the calcarine sulcus in a positive and a negative
part potentially reflecting the previously reported retinotopic organisation.
This result, therefore, indicates that we in principle can detect effects of the
y-coordinates in our measurements and that the difference in correlates for x-
and y-coordinates is not solely due to our inability of detecting y-effects.
However, my new analyses also revealed a potential issue in my previous analyses
and questions my previous results.

Flipping or not flipping before averaging sources?
--------------------------------------------------
The direction of the effects of the horizontal contrasts is opposite of the
direction of the previous correlations. It turned out that this is because I
used 'mean_flip' label mode when extracting the average signal for a brain area.
So I investigated further what mean_flip actually means and what the effects
are.

mean_flip label mode flips the sign of a source current depending on the
orientation of the underlying dipole within the cortical surface. For V1 and V2
I have extracted the flip pattern (see V1V2_flip_pattern.h5) and verified that
sources on one side of the calcarine sulcus are flipped while sources on the
other side are not. Practically, this means that the nice top/bottom pattern for
the vertical contrast that I found will turn into a uniform pattern within V1 so
that, when you average across sources within V1 with flipping, you will find an
effect of the y-coordinate and, when you average without flipping, you will not.
So without flipping you will get the situation in V1 that we were worried about:
You will be able to see an effect of x but not of y, because you average out the
y-effect.

The problem, however, is that I shouldn't flip the source currents, because
during source reconstruction I already use fixed orientation constraints. This
means that the direction of the source currents already followed the orientation
of the cortical surface during source reconstruction such that it doesn't make
sense to flip the direction of the source currents afterwards.

While not flipping may kill the y-effect in V1 it may not do so in other areas
and we'll have to see what effects come out of the analysis. Given that similar
problems as in V1 may exist, it would perhaps be best to do the GLM directly on
individual sources and later aggregate only the magnitudes of effects across
sources of individual areas.

semi-transparent divergent colormaps in MNE and pysurfer
--------------------------------------------------------
It turns out that MNE together with pysurfer handles divergent colormaps very
badly. I needed this to display t-values for the horizontal and vertical
contrasts on the inflated brain. For standard colormaps you can choose a
mid-value until which the colormap becomes less transparent, allowing the cortex
visualisation to shine through the data in places where there are low values.
For some reason I couldn't get this to work properly with MNE's divergent
colormap. Probably that's because pysurfer doesn't provide a suitable interface
for this with a divergent colormap.

Based on pysurfer's routine for scaling a colormap to desired values I have,
therefore, created my own implementation which works nicely with the coolwarm
colormap which is much better than MNE's mne colormap.

I noticed that sometimes the light source seems to be located differently
(sometimes the left of the brain is light and sometimes the right). I wonder
whether I can control that.

Unbalanced design
-----------------
Checking the design matrix for the binary regression I noticed that there are
more dots presented in the top half of the screen than in the bottom half, but
the x-coordinates are nicely balanced. This is because Hame selected specific
trials for the experiment, but then mirrored them at the y-axis to get the same
number of left and right trials. This perfectly balanced the x-coordinates, but
doubled a potential bias of the y-coordinates. Specifically, 60% of all first
dots were in the top half of the screen in the BeeMEG experiment.

I have not yet assessed what impact this could have on the analysis, but my
intuition is that it is not too big.