Gram-Schmidt orthogonalised accumulated evidence
Stefan suggested to include an accumulated evidence regressor that is explicitly
orthogonalised with respect to the corresponding current dot x-position. That's
because he didn't trust my accev regressor which is simply the accumulated
evidence up to the previous dot. So I implemented formal orthogonalisation of
the log-posterior odds (accumulated evidence) with respect to dot_x. It turns
out, however, that the resulting regressor is very strongly correlated with
accev (correlation magnitude > 0.99). So using formal orthogonalisation just
recovers the sum of dot_x up to the previous dot. It also has the disadvantage
that the signs of the resulting regressor may be arbitrary. For example, I see
that accev and lpr_orth are strongly anti-correlated for the 3rd dot for some
subjects and strongly correlated for other subjects.

These relationships should not have surprised me in hindsight given that
everything is linear and the only difference between dot_x and accumulated
evidence is that the current dot_x is added to the previous accumulated
evidence. Given the uninterpretability of the sign of lpr_orth, it's certainly a
worse regressor than accev.