checking relation between accev, motor preparation and dot counts
I investigated the influence of possible confounds on our analysis of
accumulated evidence. This was triggered by Stefan's worry that accumulated
evidence should correlate with any signal that rises within trial. What he
forgot in that thinking is that we look for correlations across trials for which
the idea of a general increase within trial is not necessarily influential. In
the analysis, I, consequently, show that a motor preparation regressor, which
rises indiscriminately towards the response time, is only weakly correlated with
accev (magnitude < 0.1 on average across subjects). Because dot counts do not
vary across trials at all, they are pretty much unrelated to accev.