analysis with new normalisation and motor preparation (motoprep) regressor
Given my latest insights about non-confounding motor preparation signals, I
decided to include the corresponding motor preparation regressor into the
analysis. At the same time, I have updated the normalisation so that it better
maintains the assumptions about the underlying quantities and I now explicitly
check whether the design matrix is collinear. For the present analysis I have
also excluded the regressors that we are not really interested in (accsur_pca,
dot_x_cflip, accev_cflip, move_dist).

Including the motoprep regressor meant that I had to allow time-dependent design
matrices. Also, I decided to do normalisation of motoprep across all times so
that motoprep has a global scale. This means, that the estimated betas should in
the ideal case be equal across time points, but the variance of the estimate
should still depend on how many trials contribute to the estimation of motoprep
at a given time t0.

normalisation
-------------
In the new normalisation I now decided to not centre the regressors unless this
would introduce strong collinearity with the intercept. This would happen, if
the mean of the regressor is considerably higher than its standard deviation.
I'm not 100% certain of what I want to achieve with this. Not centering should
not affect the beta estimates of the regressors except for that of the
intercept which will loose the part corresponding to the mean of the regressor,
as the regressor now represents its mean itself, i.e., I model the means of the
regressors so that the intercept only contains unexplained/unexpected average
effects. Only the entropy-regressor has to be centred to avoid collinearity with
the intercept. All other positive-only regressors (e.g., abs_dot_x and
especially motoprep) will now reduce the intercept effect by postulating
either a consistent positive or negative change in the measured signal.

I also came up with a new scaling for the potentially non-centred regressors.
This is motivated by the procedures for diagnosing collinearity described in
Rawlings, Pantula and Dickey (1998) where regressors, as columns in a design
matrix, are scaled to have unit vector length. Because the resulting scale of
the individual elements in the regressor would depend on their count, i.e., the
number of data points, I divide by the vector length, but then multiply by the
square root of the number of elements. This recovers that the elements of the
intercept are 1. Also, the resulting standard deviations of balanced regressors
with switching signs are close to 1, implementing a similar normalisation as
before. Only for the regressors with consistent sign the standard deviation will
be below 1. This is because, taking dot_x and abs_dot_x as examples, both would
be normalised by the same scaling such that the magnitudes of their absolute
values are matched, but if signs are taken into account, the spread of dot_x is
double that of abs_dot_x such that after normalisation the standard deviation of
abs_dot_x is only 0.5. I could also go back to normalising the spread of the
regressor instead of the magnitude of the absolute of its elements. My main
motivation to do normalisation is to have beta-estimates with comparable scale
so that I can apply the same significance criterion to all of them (cf. the
p_thresh parameter). Well, one could argue that dot_x and abs_dot_x do have a
common scale on the level of individual elements which is pixel and therefore I
should not normalise their spread, as this would remove a physical
interpretation. In the end this could be important for which temporal clusters I
detect as significant, because I so far create an empirical distribution of
effects and compute p-values based on its quantiles. If certain regressors have
consistently lower beta-estimates than others, this would disadvantage them in
the selection of significant temporal clusters. I should further investigate and
make a final decision!

collinearity diagnosis
----------------------
I now compute condition numbers of the design matrices to explicitly detect
possible collinearities. The results are stored in the results HDF store under
DM_condition_numbers. Values below 10 are good (Rawlings et al., 1998). For the
committed choice of regressors the condition numbers are above 10 for some
subjects. This appears to be caused by the low number of trials contributing
to motoprep at early time points and, especially, for subjects with quite late
responses.

results of the analysis are currently computed and will be found in
	source_sequential_201709011135.h5