bugfix + trialregs_dot=0 + normalising standard deviations
normalisation
-------------
I have decided to switch back to scaling regressors so that their standard
deviation equals 1. The main reason is comparability in the resulting
beta-estimates. In particular, I have confirmed that the beta of a regressor
halves when the scaling of the regressor doubles. I was worried that as a result
of the vector-based normalisation the unsigned regressors (e.g., abs_dot_x)
consistently have higher betas than the signed ones, because their standard
deviation is only around 0.5 while it is around 1 for the signed regressors. It
may be the case that the vector-based normalisation makes more sense with
respect to the underlying physical units, but when I want to compare the sizes
of effects across regressors, I think the normalisation with respect to
standard deviation makes more sense. At least then I have a common statistical
unit.

trialregs_dot
-------------
During these considerations I also concluded that it doesn't make sense at all
to set parts of the trial regressors to 0. Already for the last analysis, when
thinking about sensible normalisation of regressors, I thought that it's not the
greatest idea to include trial regressors in the dot-based sequential analysis,
because they cannot exhibit any sequential effects.

When normalising the design matrix such that the stds are 1, I noticed a strange
effect of trial_time: I expected that the std of trial regressors is below 1
after I set values for 2 of the three dots to 0, but trial_time had std greater
than 1. This was because trial_time defines a uniform distribution, but adding
probability mass to one of the edges of the distribution does apparently not
change the std much. I didn't expect that. In any case, it tipped the scales for
me towards thinking that setting values of trial regressors to 0 doesn't make
sense, because that introduced strong differences in the stds of the trial
regressors (response and entropy close to 0.57, trial_time close to 1) which
would have got me in the situation that I wanted to avoid by normalising
regressors.

The cleanest interpretation of trial regressors comes simply from a standard,
non-sequential analysis, but I left them in now with trialregs_dot=0. The
effects can then be interpreted in the sense that perhaps also trial regressors
are represented by the brain at a specific time after dot onset, i.e., in a
dot-specific manner. Put differently, the trial regressors in the sequential
analysis represent effects that return with the same magnitude or rather
direction at a fixed delay after dot onset, i.e., every 100 ms. If, however, the
true effects of trial regressors occur only at a single time point within the
trial, I would, given that they are strong enough, still see a diminished
representation of them in this analysis.

bugfix in condition number computation
--------------------------------------
I did not properly normalise columns of the design matrix in subject_DM.cnumfun,
because I forgot to tell numpy over which axis to sum so that it summed over the
whole matrix. I only noticed this when I switched to std normalisation of
regressors, because this lead to increased condition numbers, even though
different scaling shouldn't have lead to changes in condition numbers.

new analysis with the made changes is running and will be available in:
	source_sequential_201709011758.h5

and corresponding *slabs files.