investigating the motor preparation signal and other time-indexed regressors
Checking results from the new sequential analysis with motoprep regressor I
realised that the sequential analysis is not very good in producing reliable
results for the time-indexed regressors such as motoprep. This is because at any
given time point within the trial information from these regressors is only
partially represented. Specifically for motoprep in which most elements are 0
for early time points the effects are driven by very few data points, but the
standard errors for the estimated betas are still very low, because it is
assumed that all trials contribute to the computation of the betas. My first
intuition was to enhance the statistical analysis so that it better reflects the
effective number of trials for an effect, but this is hard to define. So I
eventually decided that it would be cleanest to make an analysis which includes
all available time points of all trials such that all trials more or less
equally contribute to the regression. The resulting analysis parallels the
typical fMRI analyses where you put a response function (here: motor preparation
signal, there: hemodynamic response) onto an event (here: response time, there:
some stimulus event or whatever). The instructions for the analysis are in
infer_time_source.py and results can be viewed using show_labels_time.py

I also included two additional time-indexed regressors into the analysis:
dotcount which is a step-signal that rises every 100 ms and stays equal after
the response and motoresponse which is the motor preparation signal, but with
sign according to which choice the subject eventually made. I verified that the
resulting design matrix has no collinearities. On the second level I simply do a
t-test to check for signals that are consistent across subjects. I correct for
multiple comparison using FDR correction across all three regressors.

Results
-------
can be computed quickly (5 min), but are also stored in
	source_time_201709051204.h5

an earlier run only with motoprep is stored in
	source_time_201709041745.h5

motoprep
........
The only areas surviving multiple comparison correction are bilateral parts of
the anterior cingulate (areas 24, 32), right dorsolateral prefrontal cortex and
right area 5L. Checking the t-values of all areas you can see that there are
also weaker effects in motor areas and temporal areas. I wonder why I find
specifically for the indiscriminate rise (and fall) of currents towards a
response effects in dorsolateral prefrontal cortex?

motoresponse
............
This regressor looks for signals which rise and fall depending on the choice of
a subject. Compared to the other regressors it has the largest effects with
t-values above +/-7. These highest t-values are only reached by bilateral
primary motor cortices and parts of right area 24 (cingulate). In total,
however, there are 54 areas with significant motoresponse effects. These include
large parts of the somatosensory and motor system and the cingulate, but also
visual areas (eg. V3, V6, V7), parietal (PFt, 7PC, AIP, LIPd) and auditory
areas (A1, LBelt). As far as I can tell from a quick visual check, the largest
effects overlap mostly with those from the response regressor in the sequential
analysis verifying the results there. Yet, the broad activations I see for
motoresponse, which, for example, also include bilateral areas v23ab, suggest
that some effects of the response and accumulated evidence overlap. This is not
so surprising, because seen across the whole trial an area which represents
accumulated evidence should also eventually represent the response, especially
around the time of response. However, I should be able to differentiate
accumulated evidence early in the trial as it fluctuates more independently of
the final response.

dotcount
........
looks for signals that continuously rise or fall within a trial from the moment
of the first dot onset and plateau after the response. As many responses are
made after 0.9 s the plateau should not contribute that much. Also it probably
doesn't matter much that the rise or fall happens in 100 ms steps. This
regressor has the weakest effects with only 4 of them being significant after
mutliple comparison correction. These are only in the left hemisphere and in
premotor eye areas (FEF, PEF), inferior frontal cortex (IFJa) and frontal
opercular cortex (FOP4). I wonder whether this is meaningful. Below threshold
more widespread effects in insular/opercular, motor, temporal, anterior
cingulate and frontal areas can be seen.

intercept
.........
identifies signals that are consistently different from 0 on average across all
trials and time points (after baseline correction in all trials). Effects are
widespread, especially occurring in bilateral insula and operculum,
somatosensory and motor areas, temporal cortex and left dorsolateral prefrontal
cortex. I guess these effects could mean anything.

Caveats
-------
I also computed permutations. A first check showed that there are similarly many
significant effects after permutation as before permutation. This is not what I
expected, because I assumed that permutation would break effects easily and
produce only unreliable results. Apparently this is not the case and I should
check what's going on. Or are motoprep effects really that small and unreliable?