checking effect of delay on results for accev_time regressor
Stefan suggested to do this and it turns out to be good for excluding the
potential confound between accumulated evidence and response, especially with
respect to effects in motor cortex. See accev_response_correlations.ipynb for a
more detailed discussion of the issue.

New results are stored in:
	source_time_201709071202.h5 (delay 0)
	source_time_201709071302.h5 (delay 0.2)
	source_time_201709071413.h5 (delay 0.4)

It appears to me that the delay of 0.3 actually has the largest effects with
clearly diminishing effects already for 0.2 and 0.4, but I haven't formally
looked at this. Specifically, motor effects also disappear for small delays even
though the correlation between accev_time and motoresponse is larger then.

The permutations now show no significant results anymore, for no regressor, as
far as I can remember. This is great as it supports the original results. I
wonder why I saw significant results after permutation when I first looked at
this? Perhaps this does not happen anymore, because I now correct for multiple
comparisons across all regressors.