new analysis including first 5 dots again
Stefan wanted to see results for the first 5 dots, probably because accumulated
evidence for the first 3 dots is only based on the first 2 dots and he wanted
more power, but I don't remember properly anymore. I decided that there is no
point in including any trial regressors in the analysis, as they are anyway hard
to interpret and I planned a singledot analysis as well. As far as I can tell
from a first glance, the results for the first 5 dots are quite similar to those
for the first 3 except that they are perhaps a bit less noisy.

I have the impression that these results (3 and 5 dots) show fewer effects in
superior parietal cortex than the previous analyses that I reported in the last
version of the manuscript, but I haven't checked that thoroughly. Anyway, one
question about the results is why we don't see many/strong effects in superior
parietal cortex for accumulated evidence? After all, this is the hub that
contains LIP that most monkey studies concentrate on and it supposedly is a
multimodal region in which visual and somatosensory information come together.
But I have also not thoroughly investigated this, yet.

Results are stored in:
	source_sequential_201709061827.h5